+++ libraries/dagster-openlineage/dagster_openlineage_tests/test_adapter_step.py
@@ -11,7 +11,7 @@ from openlineage.client.uuid import generate_new_uuid
 
 from dagster_openlineage.adapter import OpenLineageAdapter
 
-from .conftest import PRODUCER
+from .conftest import DEFAULT_JOB_FACETS, PRODUCER
 
 
 def _expected_parent_facet(pipeline_run_id: str, pipeline_name: str):
@@ -50,7 +50,7 @@ def test_start_pipeline(mock_client, mock_to_utc_iso_8601):
             job=Job(
                 namespace=DEFAULT_NAMESPACE_NAME,
                 name=f"{pipeline_name}.{step_key}",
-                facets={},
+                facets=DEFAULT_JOB_FACETS,
             ),
             producer=PRODUCER,
             inputs=[],
@@ -90,7 +90,7 @@ def test_complete_step(mock_client, mock_to_utc_iso_8601):
             job=Job(
                 namespace=DEFAULT_NAMESPACE_NAME,
                 name=f"{pipeline_name}.{step_key}",
-                facets={},
+                facets=DEFAULT_JOB_FACETS,
             ),
             producer=PRODUCER,
             inputs=[],
@@ -128,7 +128,7 @@ def test_fail_step(mock_client, mock_to_utc_iso_8601):
             job=Job(
                 namespace=DEFAULT_NAMESPACE_NAME,
                 name=f"{pipeline_name}.{step_key}",
-                facets={},
+                facets=DEFAULT_JOB_FACETS,
             ),
             producer=PRODUCER,
             inputs=[],